+++ src/QuickMarcEditor/useSaveRecord/useSaveRecord.js
@@ -9,7 +9,6 @@ import { useStripes } from '@folio/stripes/core';
 
 import {
   useAuthorityLinking,
-  useIsShared,
   useValidation,
 } from '../../hooks';
 import { QUICK_MARC_ACTIONS } from '../constants';
@@ -40,6 +39,7 @@ const useSaveRecord = ({
   refreshPageData,
   onClose,
   onSave,
+  onCreateAndKeepEditing,
 }) => {
   const stripes = useStripes();
 
@@ -48,13 +48,13 @@ const useSaveRecord = ({
     marcType,
     initialValues,
     instance,
+    isShared,
   } = useContext(QuickMarcContext);
   const {
     linkableBibFields,
     linkingRules,
     sourceFiles,
   } = useAuthorityLinking({ marcType, action });
-  const { isShared } = useIsShared();
 
   const isRequestToCentralTenantFromMember = applyCentralTenantInHeaders(isShared, stripes, marcType);
   const centralTenantId = stripes.user.user.consortium?.centralTenantId;
@@ -155,6 +155,7 @@ const useSaveRecord = ({
     refreshPageData,
     onClose,
     onSave,
+    onCreateAndKeepEditing,
   });
 
   const runValidation = useCallback(async (formValues) => {