+++ src/QuickMarcEditor/useSaveRecord/useSubmitRecord/useSumbitRecord.js
@@ -29,10 +29,7 @@ import {
 } from '../../utils';
 import getQuickMarcRecordStatus from '../../getQuickMarcRecordStatus';
 import { QuickMarcContext } from '../../../contexts';
-import {
-  useAuthorityLinking,
-  useIsShared,
-} from '../../../hooks';
+import { useAuthorityLinking } from '../../../hooks';
 import { useMarcRecordMutation } from '../../../queries';
 import { QUICK_MARC_ACTIONS } from '../../constants';
 
@@ -45,6 +42,7 @@ const useSubmitRecord = ({
   refreshPageData,
   onClose,
   onSave,
+  onCreateAndKeepEditing,
 }) => {
   const {
     externalId: _externalId,
@@ -66,11 +64,12 @@ const useSubmitRecord = ({
     instance,
     continueAfterSave,
     relatedRecordVersion,
+    setIsShared,
+    isUsingRouter,
   } = useContext(QuickMarcContext);
 
   const { actualizeLinks } = useAuthorityLinking({ marcType, action });
   const { updateMarcRecord } = useMarcRecordMutation({ tenantId });
-  const { setIsShared } = useIsShared();
 
   const redirectToRecord = useCallback(async (externalId, instanceId) => {
     if (marcType === MARC_TYPES.HOLDINGS) {
@@ -96,21 +95,35 @@ const useSubmitRecord = ({
       setIsShared(false);
     }
 
-    const routes = {
-      [MARC_TYPES.BIB]: `${basePath}/edit-bibliographic/${externalId}`,
-      [MARC_TYPES.AUTHORITY]: `${basePath}/edit-authority/${externalId}`,
-      [MARC_TYPES.HOLDINGS]: `${basePath}/edit-holdings/${externalId}`,
-    };
+    if (isUsingRouter) {
+      const routes = {
+        [MARC_TYPES.BIB]: `${basePath}/edit-bibliographic/${externalId}`,
+        [MARC_TYPES.AUTHORITY]: `${basePath}/edit-authority/${externalId}`,
+        [MARC_TYPES.HOLDINGS]: `${basePath}/edit-holdings/${externalId}`,
+      };
 
-    // use `history.location.search` instead of `location.search` because `setIsShared` also
-    // sets `shared` url parameter so we need to keep it here without overriding
-    history.push({
-      pathname: routes[marcType],
-      search: history.location.search,
-    });
+      // use `history.location.search` instead of `location.search` because `setIsShared` also
+      // sets `shared` url parameter so we need to keep it here without overriding
+      history.push({
+        pathname: routes[marcType],
+        search: history.location.search,
+      });
+    } else {
+      onCreateAndKeepEditing(externalId);
+    }
 
     await refreshPageData(fieldIds, QUICK_MARC_ACTIONS.EDIT, externalId);
-  }, [basePath, marcType, history, refreshPageData, stripes, action, setIsShared]);
+  }, [
+    basePath,
+    marcType,
+    history,
+    refreshPageData,
+    stripes,
+    action,
+    setIsShared,
+    isUsingRouter,
+    onCreateAndKeepEditing,
+  ]);
 
   const onCreate = useCallback(async (formValues, _api) => {
     const formValuesToProcess = prepareForSubmit(formValues);